+++ src/asbflow/auth/__init__.py
@@ -1,7 +1,9 @@
 ﻿from asbflow.auth.base import ASBClientProvider
 from asbflow.auth.factory import ASBClientProviderFactory
 from asbflow.auth.providers import (
+    ClientSecretCredentialClientProvider,
     ConnectionStringClientProvider,
+    DefaultAzureCredentialClientProvider,
     ManagedIdentityClientProvider,
 )
 
@@ -10,4 +12,6 @@ __all__ = [
     "ASBClientProviderFactory",
     "ConnectionStringClientProvider",
     "ManagedIdentityClientProvider",
+    "DefaultAzureCredentialClientProvider",
+    "ClientSecretCredentialClientProvider",
 ]